fix some build errors and signed/unsigned issues. also: don't forget to have virtual deconstructors when you plan to derive from a class!

+++ otserv/trunk/account.h
@@ -37,7 +37,7 @@ public:
   Account();
   ~Account();
 
-  int accnumber;
+  unsigned long accnumber;
   int accType;     // ?
   int premDays;    // Premium days
